# Optimize checksum computation