--- a/docs/changes/io.fits/17205.feature.rst
+++ b/docs/changes/io.fits/17205.feature.rst
@@ -1,0 +1,1 @@
+Optimize checksum computation.