@mhvk 2024-10-18
Hmm, this whole routine felt like an example of coding when one does not have access to `long int` - and the whole lot can be replaced by a much simpler summation -- will push PR shortly...