@pllim 2024-10-18
Hmm regression under certain condition?

```
    def test_failing_compressed_datasum(self):
        """
        Regression test for https://github.com/astropy/astropy/issues/4587
        """
        n = np.ones((10, 10), dtype="float32")
        comp_hdu = fits.CompImageHDU(n)
        comp_hdu.writeto(self.temp("tmp.fits"), checksum=True)
    
>       with fits.open(self.temp("tmp.fits"), checksum=True) as hdul:
...
astropy.utils.exceptions.AstropyUserWarning: Checksum verification failed for HDU ('COMPRESSED_IMAGE', 1).
```

The other one seems to come from this in `test_connect.py` with `<_io.FileIO [closed]>`:

```python
    def test_memmap(self, tmp_path):
        filename = tmp_path / "test_simple.fts"
        t1 = Table(self.data)
        t1.write(filename, overwrite=True)
        t2 = Table.read(filename, memmap=False)
        t3 = Table.read(filename, memmap=True)
        assert equal_data(t2, t3)
        # To avoid issues with open files, we need to remove references to
        # data that uses memory mapping and force the garbage collection
        del t1, t2, t3
        gc.collect()
```

Anyway, I restarted the failing "double run" job, just in case.