@pllim 2024-10-18
Also the double run job uses the test runner we been thinking about getting rid of...

`double: python -c 'import sys; from astropy import test; test(); sys.exit(test())'`