@mhvk 2024-10-18
See #17209 for my alternative.